JS fixes and nit cleanup
